% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_assumptions}
\alias{check_assumptions}
\title{check_assumptions function
checking main assumptions of statistical tests or reporting the results of the implicit checks of R functions}
\usage{
check_assumptions(df, assumption, test_name)
}
\arguments{
\item{df}{input data}

\item{assumption}{the name of the assumption of the test
param test_name the name of the statistical test}
}
\description{
check_assumptions function
checking main assumptions of statistical tests or reporting the results of the implicit checks of R functions
}
\examples{
check_assumptions(df = ggplot2::diamonds, assumption = ..., test_name= ...)

this is still a placeholder, under construction
}
