% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_assumptions_data}
\alias{check_assumptions_data}
\title{check_assumptions functions:
checking main assumptions about data:
made by the data distribution itself, at a different sampling point in time
or by using tests/models}
\usage{
check_assumptions_data(df1, df2)
}
\arguments{
\item{df1}{input data}

\item{df2}{input data}

\item{df}{input data}

\item{assumption}{the name of the assumption of the test}

\item{test_name}{the name of the statistical test}

\item{model_name}{the name of the statistical model}
}
\description{
check_assumptions functions:
checking main assumptions about data:
made by the data distribution itself, at a different sampling point in time
or by using tests/models
}
\examples{
check_assumptions_data(df1 = iris[1:50,]$Petal.Length, df2= iris[101:150,]$Petal.Length)
}
