% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_assumptions_test}
\alias{check_assumptions_test}
\title{check_assumptions function
checking main assumptions of statistical tests or reporting the results of the implicit checks of R functions}
\usage{
check_assumptions_test(df, test_name, assumption)
}
\arguments{
\item{df}{input data}

\item{test_name}{the name of the statistical test}

\item{assumption}{the name of the assumption of the test}

\item{model_name}{the name of the statistical model}
}
\description{
check_assumptions function
checking main assumptions of statistical tests or reporting the results of the implicit checks of R functions
}
\examples{
check_assumptions_test(df = ggplot2::diamonds, assumption = aa, test_name= tt)

this is still under construction
}
