% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_ts.R
\name{f1}
\alias{f1}
\title{seasonality complete analysis, see Robert Hyndamnn, as at:
https://robjhyndman.com/hyndsight/tscharacteristics/}
\usage{
f1(x, a, b)
}
\description{
seasonality complete analysis, see Robert Hyndamnn, as at:
https://robjhyndman.com/hyndsight/tscharacteristics/
}
