% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_ts.R
\name{rev_ts_multiv}
\alias{rev_ts_multiv}
\title{multivariate ts
using Hyndman´s anomalous and tsfeatures packages --------------
https://robjhyndman.com/hyndsight/tscharacteristics/}
\usage{
rev_ts_multiv(tsmultiv, ...)
}
\description{
multivariate ts
using Hyndman´s anomalous and tsfeatures packages --------------
https://robjhyndman.com/hyndsight/tscharacteristics/
}
