% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_ts.R
\name{rev_ts_univ}
\alias{rev_ts_univ}
\title{rev_ts function
 for main diagnostics on general time series data
auto-correlation, cross-correlation, stationarity tests,
seasonality and related measures}
\usage{
rev_ts_univ(tsuniv, ...)
}
\description{
rev_ts function
 for main diagnostics on general time series data
auto-correlation, cross-correlation, stationarity tests,
seasonality and related measures
}
