% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_variability.R
\name{rev_variability}
\alias{rev_variability}
\title{rev_variability function for general data sets}
\usage{
rev_variability(df, ...)
}
\arguments{
\item{df}{input data
...other important comments}
}
\description{
rev_variability function for general data sets
}
\examples{
rev_variability(df=ggplot2::diamonds)

}
