% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revaliew.r
\docType{package}
\name{revaliew -package}
\alias{revaliew -package}
\alias{revaliew}
\title{revaliew}
\description{
Advanced data validation and review made easy
}
\details{
Built on a large number of reliable R-packages
with the purpose of automating important tasks like:
describing the main uni and multi-variate characteristics of a data set;
finding anomalous patterns, clusters and/or outliers in data;
verifying assumptions about data (and models/statistical tests);
evaluating glm or time series models;
reporting using quantitative measures and visual representations.
}
