% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewd.R
\name{reviewd}
\alias{reviewd}
\title{fast reporting results function
calling all functions or as needed}
\usage{
reviewd(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
fast reporting results function
calling all functions or as needed
}
\examples{
reviewd(df = ggplot2::diamonds)
}
