% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_assoc.R
\name{view_assoc}
\alias{view_assoc}
\title{view association rules}
\usage{
view_assoc(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view association rules
}
\examples{
view_assoc(df = datasets::iris)
}
