% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_clusters.R
\name{view_clusters}
\alias{view_clusters}
\title{view_clusters function}
\usage{
view_clusters(df)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view_clusters function
}
\examples{
view_clusters(df = iris[,1:4])
}
