% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data.R
\name{view_data}
\alias{view_data}
\title{view_data function
short description of data: only summaries and summarizing plots}
\usage{
view_data(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view_data function
short description of data: only summaries and summarizing plots
}
\examples{
view_data(df = ggplot2::diamonds)


}
