% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_outliers.R
\name{view_outliers}
\alias{view_outliers}
\title{view_outliers function
plots and scores for outlier detection}
\usage{
view_outliers(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view_outliers function
plots and scores for outlier detection
}
\examples{
view_data(df = ggplot2::diamonds)


}
