% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_model.R
\name{rev_model}
\alias{rev_model}
\title{rev_model function
 for main diagnostics on goodness of fit
 residual testing}
\usage{
rev_model(mts, model_a, ...)
}
\arguments{
\item{mts}{mts=m means glm or lm model, mts=ts means time series model}
}
\description{
rev_model function
 for main diagnostics on goodness of fit
 residual testing
}
