% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_variability.R
\name{rev_variability}
\alias{rev_variability}
\title{the basic methods of resampling with replacement, paramteric and nonparametric, are general enough to cover
main statistics for general data sets and basic modeling formulae
with confidence intervals and corresponding plots
for details one is referred to the boot package manual}
\usage{
rev_variability(df, ...)
}
\description{
the basic methods of resampling with replacement, paramteric and nonparametric, are general enough to cover
main statistics for general data sets and basic modeling formulae
with confidence intervals and corresponding plots
for details one is referred to the boot package manual
}
