% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_multivar.R
\name{view_multivar}
\alias{view_multivar}
\title{view_multivar function
multivariate analysis and views: pairwise plots, correlations, cumulative distributions}
\usage{
view_multivar(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view_multivar function
multivariate analysis and views: pairwise plots, correlations, cumulative distributions
}
\examples{
overview_multivar(df = ggplot2::diamonds)


}
