% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_univar.R
\name{view_univar}
\alias{view_univar}
\title{view_univar function
univariate analysis: distribution functions}
\usage{
view_univar(df, ...)
}
\arguments{
\item{df}{input data
....other important comments}
}
\description{
view_univar function
univariate analysis: distribution functions
}
\examples{
view_univar(df = ggplot2::diamonds)
}
